refactor(mcp): colocate management tool schemas
