+++ packages/server/lib/controllers/mcp/integrations/list.ts
@@ -1,13 +1,9 @@
-import * as z from 'zod/v4';
-
 import integrationService from '../../../services/integration.service.js';
 import { defineManagementMcpTool } from '../managementTool.js';
 import { integrationToMcp } from './formatter.js';
-import { listIntegrationsOutputSchema } from './schema.js';
+import { listIntegrationsArgumentsSchema, listIntegrationsOutputSchema } from './schema.js';
 
 import type { ListIntegrationsOutput } from './schema.js';
-
-const listIntegrationsArgumentsSchema = z.object({}).strict();
 
 export const listIntegrationsTool = defineManagementMcpTool<typeof listIntegrationsArgumentsSchema, ListIntegrationsOutput>({
     name: 'integrations_list',
