+++ packages/server/lib/controllers/mcp/integrations/schema.ts
@@ -1,6 +1,96 @@
 import * as z from 'zod/v4';
 
-const integrationCredentialsSchema = z.discriminatedUnion('type', [
+import {
+    integrationCredentialsSchema,
+    integrationDisplayNameSchema,
+    integrationForwardWebhooksSchema,
+    providerConfigKeySchema,
+    providerSchema
+} from '../../../helpers/validation.js';
+
+const createIntegrationBaseArguments = {
+    provider: providerSchema,
+    integration_id: providerConfigKeySchema,
+    display_name: integrationDisplayNameSchema,
+    forward_webhooks: integrationForwardWebhooksSchema
+};
+
+// The MCP SDK only advertises top-level Zod object schemas; a discriminated union is emitted as an empty object schema.
+// Keep this as an object and use metadata to expose the conditional fields to clients while superRefine enforces them at runtime.
+export const createIntegrationArgumentsSchema = z
+    .object({
+        ...createIntegrationBaseArguments,
+        credential_source: z.enum(['nango', 'own']).describe('Use nango for Nango-provided credentials or own for caller-supplied credentials.'),
+        credentials: integrationCredentialsSchema.optional().describe('Only applicable when credential_source is own.'),
+        integration_config: z.record(z.string(), z.string().max(8192)).optional().describe('Only applicable when credential_source is own.')
+    })
+    .strict()
+    .superRefine((args, ctx) => {
+        if (args.credential_source !== 'nango') {
+            return;
+        }
+
+        if (args.credentials !== undefined) {
+            ctx.addIssue({
+                code: 'custom',
+                path: ['credentials'],
+                message: 'credentials is only allowed when credential_source is own'
+            });
+        }
+
+        if (args.integration_config !== undefined) {
+            ctx.addIssue({
+                code: 'custom',
+                path: ['integration_config'],
+                message: 'integration_config is only allowed when credential_source is own'
+            });
+        }
+    })
+    .meta({
+        oneOf: [
+            {
+                properties: { credential_source: { const: 'nango' } },
+                not: {
+                    anyOf: [{ required: ['credentials'] }, { required: ['integration_config'] }]
+                }
+            },
+            {
+                properties: { credential_source: { const: 'own' } }
+            }
+        ]
+    });
+
+export const getIntegrationArgumentsSchema = z
+    .object({
+        integration_id: providerConfigKeySchema,
+        include: z
+            .array(z.enum(['webhook', 'credentials']))
+            .max(2)
+            .optional()
+    })
+    .strict();
+
+export const listIntegrationsArgumentsSchema = z.object({}).strict();
+
+export const updateIntegrationsArgumentsSchema = z
+    .object({
+        integration_id: providerConfigKeySchema,
+        new_integration_id: providerConfigKeySchema.optional(),
+        display_name: integrationDisplayNameSchema,
+        credentials: integrationCredentialsSchema.optional(),
+        forward_webhooks: integrationForwardWebhooksSchema,
+        integration_config: z.record(z.string(), z.string().max(8192)).optional(),
+        custom: z.record(z.string(), z.string()).optional()
+    })
+    .strict();
+
+export const deleteIntegrationsArgumentsSchema = z
+    .object({
+        integration_id: providerConfigKeySchema
+    })
+    .strict();
+
+const mcpIntegrationCredentialsSchema = z.discriminatedUnion('type', [
     z
         .object({
             type: z.enum(['OAUTH1', 'OAUTH2', 'TBA']),
@@ -54,7 +144,7 @@ export const getIntegrationOutputSchema = z
     .object({
         data: mcpIntegrationSchema.extend({
             webhook_url: z.string().nullable().optional(),
-            credentials: integrationCredentialsSchema.nullable().optional()
+            credentials: mcpIntegrationCredentialsSchema.nullable().optional()
         })
     })
     .strict();
