+++ packages/server/lib/controllers/mcp/logs/listOperations.ts
@@ -1,171 +1,15 @@
-import * as z from 'zod/v4';
-
 import { LogsDisabledError, logsOperationsService } from '@nangohq/logs';
 
 import { defineManagementMcpTool } from '../managementTool.js';
 import { PublicMcpError } from '../utils.js';
-import { defaultLimit, logsReadScope, maxLimit, normalizePeriod, periodSchema } from './utils.js';
+import { listOperationsArgumentsSchema, listOperationsOutputSchema } from './schema.js';
+import { logsReadScope, normalizePeriod } from './utils.js';
 
+import type { ListOperationsArguments } from './schema.js';
 import type { ListLogOperationsParams, ListLogOperationsResult } from '@nangohq/logs';
-import type {
-    OperationAction,
-    OperationAdmin,
-    OperationAuth,
-    OperationDeploy,
-    OperationOnEvents,
-    OperationProxy,
-    OperationSync,
-    OperationWebhook,
-    SearchOperationsState,
-    SearchOperationsType,
-    SearchPeriod
-} from '@nangohq/types';
+import type { SearchOperationsType, SearchPeriod } from '@nangohq/types';
 
 const defaultOperationsPeriodMs = 24 * 60 * 60 * 1000;
-
-const statesSchema = z
-    .array(z.enum(['waiting', 'running', 'success', 'failed', 'timeout', 'cancelled'] satisfies SearchOperationsState[]))
-    .max(10)
-    .optional();
-
-const actionOperationFilterSchema = z
-    .object({
-        type: z.literal('action'),
-        actions: z
-            .array(z.enum(['run'] satisfies OperationAction['action'][]))
-            .min(1)
-            .max(1)
-            .optional()
-    })
-    .strict();
-
-const syncOperationFilterSchema = z
-    .object({
-        type: z.literal('sync'),
-        actions: z
-            .array(
-                z.enum([
-                    'pause',
-                    'unpause',
-                    'run',
-                    'request_run',
-                    'request_run_full',
-                    'cancel',
-                    'init',
-                    'create_variant',
-                    'delete_variant'
-                ] satisfies OperationSync['action'][])
-            )
-            .min(1)
-            .max(9)
-            .optional()
-    })
-    .strict();
-
-const proxyOperationFilterSchema = z
-    .object({
-        type: z.literal('proxy'),
-        actions: z
-            .array(z.enum(['call'] satisfies OperationProxy['action'][]))
-            .min(1)
-            .max(1)
-            .optional()
-    })
-    .strict();
-
-const eventsOperationFilterSchema = z
-    .object({
-        type: z.literal('events'),
-        actions: z
-            .array(z.enum(['post_connection_creation', 'pre_connection_deletion', 'validate_connection'] satisfies OperationOnEvents['action'][]))
-            .min(1)
-            .max(3)
-            .optional()
-    })
-    .strict();
-
-const authOperationFilterSchema = z
-    .object({
-        type: z.literal('auth'),
-        actions: z
-            .array(z.enum(['create_connection', 'refresh_token', 'post_connection', 'connection_test'] satisfies OperationAuth['action'][]))
-            .min(1)
-            .max(4)
-            .optional()
-    })
-    .strict();
-
-const adminOperationFilterSchema = z
-    .object({
-        type: z.literal('admin'),
-        actions: z
-            .array(z.enum(['impersonation'] satisfies OperationAdmin['action'][]))
-            .min(1)
-            .max(1)
-            .optional()
-    })
-    .strict();
-
-const webhookOperationFilterSchema = z
-    .object({
-        type: z.literal('webhook'),
-        actions: z
-            .array(z.enum(['incoming', 'forward', 'sync', 'connection_create', 'connection_refresh'] satisfies OperationWebhook['action'][]))
-            .min(1)
-            .max(5)
-            .optional()
-    })
-    .strict();
-
-const deployOperationFilterSchema = z
-    .object({
-        type: z.literal('deploy'),
-        actions: z
-            .array(z.enum(['prebuilt', 'custom'] satisfies OperationDeploy['action'][]))
-            .min(1)
-            .max(2)
-            .optional()
-    })
-    .strict();
-
-const operationFilterSchema = z.discriminatedUnion('type', [
-    actionOperationFilterSchema,
-    syncOperationFilterSchema,
-    proxyOperationFilterSchema,
-    eventsOperationFilterSchema,
-    authOperationFilterSchema,
-    adminOperationFilterSchema,
-    webhookOperationFilterSchema,
-    deployOperationFilterSchema
-]);
-
-const listOperationsArgumentsSchema = z
-    .object({
-        search: z.string().max(256).optional(),
-        limit: z.number().int().min(1).max(maxLimit).optional().default(defaultLimit),
-        cursor: z.string().nullable().optional(),
-        states: statesSchema,
-        operations: z.array(operationFilterSchema).max(20).optional(),
-        integrations: z.array(z.string().min(1).max(256)).max(20).optional(),
-        connections: z.array(z.string().min(1).max(256)).max(20).optional(),
-        syncs: z.array(z.string().min(1).max(256)).max(20).optional(),
-        period: periodSchema.optional()
-    })
-    .strict();
-
-const listOperationsOutputSchema = z
-    .object({
-        operations: z.array(z.looseObject({})),
-        pagination: z
-            .object({
-                total: z.number(),
-                cursor: z.string().nullable()
-            })
-            .strict()
-    })
-    .strict();
-
-type ListOperationsArguments = z.infer<typeof listOperationsArgumentsSchema>;
 type ParsedListOperationsArguments = Omit<ListLogOperationsParams, 'accountId' | 'environmentId'>;
 
 export const listLogOperationsTool = defineManagementMcpTool<typeof listOperationsArgumentsSchema, ListLogOperationsResult>({
